Изменен тип данных для параметра ФОП_Количество комнат

+++ src/dosymep.Bim4Everyone/SharedParams/SharedParamsConfig.cs
@@ -329,7 +329,7 @@ namespace dosymep.Bim4Everyone.SharedParams {
         /// </summary>
         public SharedParam RoomsCount
             => new SharedParam(nameof(RoomsCount), new Guid("2498f7c7-de06-42c7-93dc-5e269cadc202")) {
-                Name = "ФОП_Количество комнат", UnitType = SharedParam.GetUnitType(nameof(RoomsCount)), StorageType = StorageType.Integer
+                Name = "ФОП_Количество комнат", UnitType = SharedParam.GetUnitType(nameof(RoomsCount)), StorageType = StorageType.Double
             };
 
         /// <summary>
